build: rebuild system test_resource_pool
