+++ backend/framework/domain/src/main/java/io/metersphere/system/domain/TestResourcePoolExample.java
@@ -564,146 +564,6 @@ public class TestResourcePoolExample {
             return (Criteria) this;
         }
 
-        public Criteria andLoadTestImageIsNull() {
-            addCriterion("load_test_image is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestImageIsNotNull() {
-            addCriterion("load_test_image is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestImageEqualTo(String value) {
-            addCriterion("load_test_image =", value, "loadTestImage");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestImageNotEqualTo(String value) {
-            addCriterion("load_test_image <>", value, "loadTestImage");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestImageGreaterThan(String value) {
-            addCriterion("load_test_image >", value, "loadTestImage");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestImageGreaterThanOrEqualTo(String value) {
-            addCriterion("load_test_image >=", value, "loadTestImage");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestImageLessThan(String value) {
-            addCriterion("load_test_image <", value, "loadTestImage");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestImageLessThanOrEqualTo(String value) {
-            addCriterion("load_test_image <=", value, "loadTestImage");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestImageLike(String value) {
-            addCriterion("load_test_image like", value, "loadTestImage");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestImageNotLike(String value) {
-            addCriterion("load_test_image not like", value, "loadTestImage");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestImageIn(List<String> values) {
-            addCriterion("load_test_image in", values, "loadTestImage");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestImageNotIn(List<String> values) {
-            addCriterion("load_test_image not in", values, "loadTestImage");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestImageBetween(String value1, String value2) {
-            addCriterion("load_test_image between", value1, value2, "loadTestImage");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestImageNotBetween(String value1, String value2) {
-            addCriterion("load_test_image not between", value1, value2, "loadTestImage");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestHeapIsNull() {
-            addCriterion("load_test_heap is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestHeapIsNotNull() {
-            addCriterion("load_test_heap is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestHeapEqualTo(String value) {
-            addCriterion("load_test_heap =", value, "loadTestHeap");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestHeapNotEqualTo(String value) {
-            addCriterion("load_test_heap <>", value, "loadTestHeap");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestHeapGreaterThan(String value) {
-            addCriterion("load_test_heap >", value, "loadTestHeap");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestHeapGreaterThanOrEqualTo(String value) {
-            addCriterion("load_test_heap >=", value, "loadTestHeap");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestHeapLessThan(String value) {
-            addCriterion("load_test_heap <", value, "loadTestHeap");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestHeapLessThanOrEqualTo(String value) {
-            addCriterion("load_test_heap <=", value, "loadTestHeap");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestHeapLike(String value) {
-            addCriterion("load_test_heap like", value, "loadTestHeap");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestHeapNotLike(String value) {
-            addCriterion("load_test_heap not like", value, "loadTestHeap");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestHeapIn(List<String> values) {
-            addCriterion("load_test_heap in", values, "loadTestHeap");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestHeapNotIn(List<String> values) {
-            addCriterion("load_test_heap not in", values, "loadTestHeap");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestHeapBetween(String value1, String value2) {
-            addCriterion("load_test_heap between", value1, value2, "loadTestHeap");
-            return (Criteria) this;
-        }
-
-        public Criteria andLoadTestHeapNotBetween(String value1, String value2) {
-            addCriterion("load_test_heap not between", value1, value2, "loadTestHeap");
-            return (Criteria) this;
-        }
-
         public Criteria andCreateUserIsNull() {
             addCriterion("create_user is null");
             return (Criteria) this;
@@ -951,76 +811,6 @@ public class TestResourcePoolExample {
 
         public Criteria andUiTestNotBetween(Boolean value1, Boolean value2) {
             addCriterion("ui_test not between", value1, value2, "uiTest");
-            return (Criteria) this;
-        }
-
-        public Criteria andGridIsNull() {
-            addCriterion("grid is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andGridIsNotNull() {
-            addCriterion("grid is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andGridEqualTo(String value) {
-            addCriterion("grid =", value, "grid");
-            return (Criteria) this;
-        }
-
-        public Criteria andGridNotEqualTo(String value) {
-            addCriterion("grid <>", value, "grid");
-            return (Criteria) this;
-        }
-
-        public Criteria andGridGreaterThan(String value) {
-            addCriterion("grid >", value, "grid");
-            return (Criteria) this;
-        }
-
-        public Criteria andGridGreaterThanOrEqualTo(String value) {
-            addCriterion("grid >=", value, "grid");
-            return (Criteria) this;
-        }
-
-        public Criteria andGridLessThan(String value) {
-            addCriterion("grid <", value, "grid");
-            return (Criteria) this;
-        }
-
-        public Criteria andGridLessThanOrEqualTo(String value) {
-            addCriterion("grid <=", value, "grid");
-            return (Criteria) this;
-        }
-
-        public Criteria andGridLike(String value) {
-            addCriterion("grid like", value, "grid");
-            return (Criteria) this;
-        }
-
-        public Criteria andGridNotLike(String value) {
-            addCriterion("grid not like", value, "grid");
-            return (Criteria) this;
-        }
-
-        public Criteria andGridIn(List<String> values) {
-            addCriterion("grid in", values, "grid");
-            return (Criteria) this;
-        }
-
-        public Criteria andGridNotIn(List<String> values) {
-            addCriterion("grid not in", values, "grid");
-            return (Criteria) this;
-        }
-
-        public Criteria andGridBetween(String value1, String value2) {
-            addCriterion("grid between", value1, value2, "grid");
-            return (Criteria) this;
-        }
-
-        public Criteria andGridNotBetween(String value1, String value2) {
-            addCriterion("grid not between", value1, value2, "grid");
             return (Criteria) this;
         }
