+++ backend/framework/domain/src/main/java/io/metersphere/system/mapper/TestResourcePoolMapper.xml
@@ -9,13 +9,10 @@
     <result column="enable" jdbcType="BIT" property="enable" />
     <result column="create_time" jdbcType="BIGINT" property="createTime" />
     <result column="update_time" jdbcType="BIGINT" property="updateTime" />
-    <result column="load_test_image" jdbcType="VARCHAR" property="loadTestImage" />
-    <result column="load_test_heap" jdbcType="VARCHAR" property="loadTestHeap" />
     <result column="create_user" jdbcType="VARCHAR" property="createUser" />
     <result column="api_test" jdbcType="BIT" property="apiTest" />
     <result column="load_test" jdbcType="BIT" property="loadTest" />
     <result column="ui_test" jdbcType="BIT" property="uiTest" />
-    <result column="grid" jdbcType="VARCHAR" property="grid" />
     <result column="server_url" jdbcType="VARCHAR" property="serverUrl" />
     <result column="deleted" jdbcType="BIT" property="deleted" />
   </resultMap>
@@ -78,8 +75,8 @@
     </where>
   </sql>
   <sql id="Base_Column_List">
-    id, `name`, `type`, description, `enable`, create_time, update_time, load_test_image, 
-    load_test_heap, create_user, api_test, load_test, ui_test, grid, server_url, deleted
+    id, `name`, `type`, description, `enable`, create_time, update_time, create_user, 
+    api_test, load_test, ui_test, server_url, deleted
   </sql>
   <select id="selectByExample" parameterType="io.metersphere.system.domain.TestResourcePoolExample" resultMap="BaseResultMap">
     select
@@ -114,15 +111,13 @@
   <insert id="insert" parameterType="io.metersphere.system.domain.TestResourcePool">
     insert into test_resource_pool (id, `name`, `type`, 
       description, `enable`, create_time, 
-      update_time, load_test_image, load_test_heap, 
-      create_user, api_test, load_test, 
-      ui_test, grid, server_url, 
+      update_time, create_user, api_test, 
+      load_test, ui_test, server_url, 
       deleted)
     values (#{id,jdbcType=VARCHAR}, #{name,jdbcType=VARCHAR}, #{type,jdbcType=VARCHAR}, 
       #{description,jdbcType=VARCHAR}, #{enable,jdbcType=BIT}, #{createTime,jdbcType=BIGINT}, 
-      #{updateTime,jdbcType=BIGINT}, #{loadTestImage,jdbcType=VARCHAR}, #{loadTestHeap,jdbcType=VARCHAR}, 
-      #{createUser,jdbcType=VARCHAR}, #{apiTest,jdbcType=BIT}, #{loadTest,jdbcType=BIT}, 
-      #{uiTest,jdbcType=BIT}, #{grid,jdbcType=VARCHAR}, #{serverUrl,jdbcType=VARCHAR}, 
+      #{updateTime,jdbcType=BIGINT}, #{createUser,jdbcType=VARCHAR}, #{apiTest,jdbcType=BIT}, 
+      #{loadTest,jdbcType=BIT}, #{uiTest,jdbcType=BIT}, #{serverUrl,jdbcType=VARCHAR}, 
       #{deleted,jdbcType=BIT})
   </insert>
   <insert id="insertSelective" parameterType="io.metersphere.system.domain.TestResourcePool">
@@ -149,12 +144,6 @@
       <if test="updateTime != null">
         update_time,
       </if>
-      <if test="loadTestImage != null">
-        load_test_image,
-      </if>
-      <if test="loadTestHeap != null">
-        load_test_heap,
-      </if>
       <if test="createUser != null">
         create_user,
       </if>
@@ -166,9 +155,6 @@
       </if>
       <if test="uiTest != null">
         ui_test,
-      </if>
-      <if test="grid != null">
-        grid,
       </if>
       <if test="serverUrl != null">
         server_url,
@@ -199,12 +185,6 @@
       <if test="updateTime != null">
         #{updateTime,jdbcType=BIGINT},
       </if>
-      <if test="loadTestImage != null">
-        #{loadTestImage,jdbcType=VARCHAR},
-      </if>
-      <if test="loadTestHeap != null">
-        #{loadTestHeap,jdbcType=VARCHAR},
-      </if>
       <if test="createUser != null">
         #{createUser,jdbcType=VARCHAR},
       </if>
@@ -216,9 +196,6 @@
       </if>
       <if test="uiTest != null">
         #{uiTest,jdbcType=BIT},
-      </if>
-      <if test="grid != null">
-        #{grid,jdbcType=VARCHAR},
       </if>
       <if test="serverUrl != null">
         #{serverUrl,jdbcType=VARCHAR},
@@ -258,12 +235,6 @@
       <if test="record.updateTime != null">
         update_time = #{record.updateTime,jdbcType=BIGINT},
       </if>
-      <if test="record.loadTestImage != null">
-        load_test_image = #{record.loadTestImage,jdbcType=VARCHAR},
-      </if>
-      <if test="record.loadTestHeap != null">
-        load_test_heap = #{record.loadTestHeap,jdbcType=VARCHAR},
-      </if>
       <if test="record.createUser != null">
         create_user = #{record.createUser,jdbcType=VARCHAR},
       </if>
@@ -275,9 +246,6 @@
       </if>
       <if test="record.uiTest != null">
         ui_test = #{record.uiTest,jdbcType=BIT},
-      </if>
-      <if test="record.grid != null">
-        grid = #{record.grid,jdbcType=VARCHAR},
       </if>
       <if test="record.serverUrl != null">
         server_url = #{record.serverUrl,jdbcType=VARCHAR},
@@ -299,13 +267,10 @@
       `enable` = #{record.enable,jdbcType=BIT},
       create_time = #{record.createTime,jdbcType=BIGINT},
       update_time = #{record.updateTime,jdbcType=BIGINT},
-      load_test_image = #{record.loadTestImage,jdbcType=VARCHAR},
-      load_test_heap = #{record.loadTestHeap,jdbcType=VARCHAR},
       create_user = #{record.createUser,jdbcType=VARCHAR},
       api_test = #{record.apiTest,jdbcType=BIT},
       load_test = #{record.loadTest,jdbcType=BIT},
       ui_test = #{record.uiTest,jdbcType=BIT},
-      grid = #{record.grid,jdbcType=VARCHAR},
       server_url = #{record.serverUrl,jdbcType=VARCHAR},
       deleted = #{record.deleted,jdbcType=BIT}
     <if test="_parameter != null">
@@ -333,12 +298,6 @@
       <if test="updateTime != null">
         update_time = #{updateTime,jdbcType=BIGINT},
       </if>
-      <if test="loadTestImage != null">
-        load_test_image = #{loadTestImage,jdbcType=VARCHAR},
-      </if>
-      <if test="loadTestHeap != null">
-        load_test_heap = #{loadTestHeap,jdbcType=VARCHAR},
-      </if>
       <if test="createUser != null">
         create_user = #{createUser,jdbcType=VARCHAR},
       </if>
@@ -350,9 +309,6 @@
       </if>
       <if test="uiTest != null">
         ui_test = #{uiTest,jdbcType=BIT},
-      </if>
-      <if test="grid != null">
-        grid = #{grid,jdbcType=VARCHAR},
       </if>
       <if test="serverUrl != null">
         server_url = #{serverUrl,jdbcType=VARCHAR},
@@ -371,13 +327,10 @@
       `enable` = #{enable,jdbcType=BIT},
       create_time = #{createTime,jdbcType=BIGINT},
       update_time = #{updateTime,jdbcType=BIGINT},
-      load_test_image = #{loadTestImage,jdbcType=VARCHAR},
-      load_test_heap = #{loadTestHeap,jdbcType=VARCHAR},
       create_user = #{createUser,jdbcType=VARCHAR},
       api_test = #{apiTest,jdbcType=BIT},
       load_test = #{loadTest,jdbcType=BIT},
       ui_test = #{uiTest,jdbcType=BIT},
-      grid = #{grid,jdbcType=VARCHAR},
       server_url = #{serverUrl,jdbcType=VARCHAR},
       deleted = #{deleted,jdbcType=BIT}
     where id = #{id,jdbcType=VARCHAR}
