+++ backend/services/system-setting/src/main/java/io/metersphere/system/service/KubernetesResourcePoolService.java
@@ -1,9 +1,9 @@
 package io.metersphere.system.service;
 
 
-import io.metersphere.system.dto.TestResourcePoolDTO;
+import io.metersphere.system.dto.TestResourceDTO;
 
 public interface KubernetesResourcePoolService {
 
-    boolean validate(TestResourcePoolDTO testResourcePool);
+    boolean validate(TestResourceDTO testResourceDT);
 }
