+++ backend/services/system-setting/src/main/java/io/metersphere/system/service/LoadResourceService.java
@@ -1,0 +1,9 @@
+package io.metersphere.system.service;
+
+import io.metersphere.system.dto.TestResourceDTO;
+
+public interface LoadResourceService {
+
+    boolean validate(TestResourceDTO testResourceDTO, String type);
+
+}
