+++ backend/services/system-setting/src/main/java/io/metersphere/system/service/NodeResourcePoolService.java
@@ -3,8 +3,9 @@ package io.metersphere.system.service;
 
 import io.metersphere.sdk.exception.MSException;
 import io.metersphere.sdk.util.Translator;
+import io.metersphere.system.domain.TestResourcePoolBlob;
+import io.metersphere.system.dto.TestResourceDTO;
 import io.metersphere.system.dto.TestResourcePoolDTO;
-import org.apache.commons.collections4.CollectionUtils;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,10 +14,11 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(rollbackFor = Exception.class)
 public class NodeResourcePoolService {
 
-    public boolean validate(TestResourcePoolDTO testResourcePool) {
-        if (CollectionUtils.isEmpty(testResourcePool.getTestResources())) {
-           throw new MSException(Translator.get("no_nodes_message"));
-        }
+    public boolean validate(TestResourceDTO testResourceDTO) {
+        /*TestResourcePoolBlob testResourcePoolBlob = testResourcePool.getConfiguration();
+        if (testResourcePoolBlob == null || testResourcePoolBlob.getConfiguration()==null){
+            throw new MSException(Translator.get("no_nodes_message"));
+        }*/
         //校验节点
         return true;
     }
