+++ backend/services/system-setting/src/main/java/io/metersphere/system/service/UiResourceService.java
@@ -1,0 +1,9 @@
+package io.metersphere.system.service;
+
+import io.metersphere.system.dto.TestResourceDTO;
+
+public interface UiResourceService {
+
+    boolean validate(TestResourceDTO testResourceDTO);
+
+}
